Revert conftest changes
